Refactor resolvers to separate file from server.js
