+++ server.js
@@ -6,9 +6,7 @@ const bodyParser = require('body-parser');
 const { graphqlExpress, graphiqlExpress } = require('apollo-server-express');
 const { makeExecutableSchema } = require('graphql-tools');
 
-
-
-
+const resolvers = require('./resolvers');
 
 const typeDefs = `
 type Cat {
@@ -24,44 +22,8 @@ type Mutation {
   removeCat(name: String!): [Cat!]!
 
 }
-
 `;
 
-
-const resolvers = {
-  Query: {
-    allCats: async (parent, args, { Cat }) => {
-      // { _id: 123123, name: "whatever"}
-      const cats = await Cat.find();
-      return cats.map((x) => {
-        x._id = x._id.toString();
-        return x;
-      });
-    },
-    oneCat: async (parent, {name}, { Cat }) => {
-      const oneKitty= await Cat.findOne({'name': name});
-
-      return oneKitty;
-    },
-  },
-  Mutation: {
-    createCat: async (parent, args, { Cat }) => {
-      // { _id: 123123, name: "whatever"}
-      const kitty = await new Cat(args).save();
-      kitty._id = kitty._id.toString();
-      return kitty;
-    },
-  removeCat: async (parent, {name}, { Cat }) => {
-    var oneKitty= await Cat.remove(Cat.findOne({'name': name}));
-    const cats = await Cat.find();
-    return cats.map((x) => {
-      x._id = x._id.toString();
-      return x;
-    });
-
-    },
-  },
-};
 const schema = makeExecutableSchema({
   typeDefs,
   resolvers,
@@ -83,10 +45,6 @@ const books = [
 ];
 */
 // The GraphQL schema in string form
-
-
-// The resolvers
-
 
 // Put together a schema
 mongoose.connect('mongodb://dbuser:cabedi1@ds127490.mlab.com:27490/ctestdb');
